==> student-performance-prediction-4/3-split-data.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split

# Memuat Dataset yang Telah Diproses
data = pd.read_csv('dataset/processed_student_data.csv')

# Pembagian Dataset
# Pisahkan fitur (X) dan target (y)
X = data.drop(columns=['exam_score'])
y = data['exam_score']

# Bagi dataset menjadi training (80%) dan testing (20%)
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

print("\n=== Ukuran Dataset ===")
print(f"Training set: {X_train.shape[0]} baris, {X_train.shape[1]} fitur")
print(f"Testing set: {X_test.shape[0]} baris, {X_test.shape[1]} fitur")

X_train.to_csv("dataset/X_train.csv", index=False)
X_test.to_csv("dataset/X_test.csv", index=False)
y_train.to_csv("dataset/y_train.csv", index=False)
y_test.to_csv("dataset/y_test.csv", index=False)
